requirements.txt
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
streamlit==1.29.0

matplotlib==3.8.2
opencv-python==4.8.1.78
scikit-learn==1.3.2